Add MPI_ANY_SOURCE test in drain_send_recv  (#126)
First, this change adds test for MPI_ANY_SOURCE in drain_send_recv_test.

Second, Libdmtcpinternal.a depends on Jbuffer() from libjalib.a and libjalib.a depends on initializeJalib() from libdmtcpinternal.a. This change fixes this inter-dependency. It calls initializeJalib in the test program so that we have the sequence
weak definition of initializeJalib in test program, real definition of initializeJalib in libdmtcpinternal.a followed by weak definition initialJalib in libjalib.a. This keeps the real definition of initializeJalib from being removed so that libjalib.a can access since initalizeJalib is used by the test program.