Fast way to find MPI_Type record to be freed
This slowness is introduced in the change of free MPI_Type records. The change goes
through all record-replay log to find the MPI_Type record. The log has
tons other MPI like MPI_Request in addition to MPI_Type. That causes slowness.
This fix goes through the log records in the reverse
order. The latest MPI_Type operations is near the end. We stop walking
through the log once we find the creator of the MPI_Type record.